second commit after finishing backed

+++ app/Http/Controllers/CategoriesController.php
@@ -17,6 +17,7 @@ class CategoriesController extends Controller
      */
     public function index()
     {
+       
         return view('categories.index')->with('categories', Category::all());
     }
 
@@ -98,6 +99,11 @@ class CategoriesController extends Controller
      */
     public function destroy(Category $category)
     {
+        if($category->posts->count() > 0)
+        {
+            session()->flash('error',"Category can not be deleted because it contain posts.");
+            return redirect()->back();
+        }
         $category->delete();
 
         session()->flash('success',"Category deleted successfully.");
